+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix `IndexError: list index out of range` error raised by `pycobertura diff` when number of files with coverage change is less then total number of files in coverage report. Thanks @ulasozguler
+
 ## 3.2.1 (2023-06-01)
 
 * Make `ZipFileSystem` read files in text mode to fix diffing mixed file systems. Thanks @ernestask